requirements.txt
asgiref==3.7.2
Django==4.2.10
django-js-asset==2.2.0
django-mptt==0.16.0
pillow==10.2.0
python-dotenv==1.0.1
sqlparse==0.4.4
tzdata==2024.1
